requirements.txt
requests
python-dotenv
pandas
pandas_ta
PyQt6
pyqtgraph